third_party/blink/web_tests/fast/forms/textarea/textarea-wrap-attribute-expected.txt

HTMLTextAreaElement.wrap reflects the wrap="" attribute.
It is not "limited to only known values", and it is a DOMString attribute which means it just returns the content attributes's value directly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

If wrap attribute is not specified it should be empty String.
PASS textArea.wrap is ''


Check if it sets warpAttr value hard, should return hard.
PASS textArea.wrap is 'hard'


Check if it sets warpAttr value as soft, should return soft.
PASS textArea.wrap is 'soft'


Check if warpAttr present but no keyVal specified, should return empty String.
PASS textArea.wrap is ''


Check if it sets warpAttr invaild value, should return foo.
PASS textArea.wrap is 'foo'


PASS successfullyParsed is true

TEST COMPLETE
